requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
kaleido==0.2.1
aiohttp==3.9.1
google-cloud-speech==2.26.0
